requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
audioop-lts==0.2.2 ; python_version >= "3.13" and python_version < "4.0"
audioread==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2026.1.4 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
decorator==4.4.2 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
imageio-ffmpeg==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
imageio==2.37.2 ; python_version >= "3.11" and python_version < "4.0"
joblib==1.5.3 ; python_version >= "3.11" and python_version < "4.0"
lazy-loader==0.4 ; python_version >= "3.11" and python_version < "4.0"
librosa==0.10.2.post1 ; python_version >= "3.11" and python_version < "4.0"
llvmlite==0.46.0 ; python_version >= "3.11" and python_version < "4.0"
moviepy==1.0.3 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.1.2 ; python_version >= "3.11" and python_version < "4.0"
numba==0.63.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==12.1.0 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.5.1 ; python_version >= "3.11" and python_version < "4.0"
pooch==1.8.2 ; python_version >= "3.11" and python_version < "4.0"
proglog==0.1.12 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and python_version < "4.0" and implementation_name != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.12.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
scikit-learn==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
scipy==1.16.3 ; python_version >= "3.11" and python_version < "4.0"
setuptools==80.9.0 ; python_version >= "3.11" and python_version < "4.0"
soundfile==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
soxr==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
standard-aifc==3.13.0 ; python_version >= "3.13" and python_version < "4.0"
standard-chunk==3.13.0 ; python_version >= "3.13" and python_version < "4.0"
standard-sunau==3.13.0 ; python_version >= "3.13" and python_version < "4.0"
threadpoolctl==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
